separation of engine logic, WIP: fix white screen
